Refactor vuex store
