+++ src/App.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import { mutations } from './store';
+import { mutations } from './store/';
 export default {
 	name: 'App',
 	computed: {
