+++ src/components/Folder.vue
@@ -38,7 +38,7 @@
 <script>
 import Vue from 'vue';
 import { Actions, ActionButton } from 'nextcloud-vue';
-import { actions, mutations } from '../store';
+import { actions, mutations } from '../store/';
 
 export default {
 	name: 'Folder',
