+++ templates/admin.php
@@ -1,6 +1,4 @@
 <?php
-/** @var $l \OCP\IL10N */
-/** @var $_ array */
 script('bookmarks', 'bookmarks.admin');
 ?>
 <div id="bookmarks" class="section"></div>
